@@ -1,17 +1,15 @@
 {
   "name": "@lab900/ui",
-  "version": "4.1.0",
+  "version": "15.0.0",
   "author": "Lab900 <info@lab900.com> (https://lab900.com)",
   "licens": "MIT",
   "sideEffects": true,
   "peerDependencies": {
-    "@angular/common": "~12.2.14",
-    "@angular/core": "~12.2.14",
-    "@angular/material": "~12.2.13",
-    "@angular/flex-layout": "^12.0.0-beta.35",
-    "@ngx-translate/core": "^13.0.0",
-    "lodash": "^4.17.20",
-    "memo-decorator": "^2.0.1"
+    "@angular/common": "~15.2.3",
+    "@angular/core": "~15.2.3",
+    "@angular/material": "~15.2.3",
+    "@ngx-translate/core": "^14.0.0",
+    "lodash": "^4.17.20"
   },
   "repository": {
     "type": "git",
