@@ -1,10 +1,10 @@
 {
   "name": "angular-library-ui",
-  "version": "1.0.0",
+  "version": "15.0.0",
   "repository": "https://github.com/lab900/angular-library-ui",
   "homepage": "https://lab900.github.io/angular-libraries",
   "author": "Lab900 <info@lab900.com> (https://lab900.com)",
-  "licens": "MIT",
+  "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve --port 4900",
@@ -34,58 +34,64 @@
     ]
   },
   "dependencies": {
-    "@angular-material-components/datetime-picker": "^6.0.3",
-    "@angular/animations": "~12.2.14",
-    "@angular/cdk": "~12.2.13",
-    "@angular/common": "~12.2.14",
-    "@angular/compiler": "~12.2.14",
-    "@angular/core": "~12.2.14",
-    "@angular/flex-layout": "^12.0.0-beta.35",
-    "@angular/forms": "~12.2.14",
-    "@angular/material": "~12.2.13",
-    "@angular/platform-browser": "~12.2.14",
-    "@angular/platform-browser-dynamic": "~12.2.14",
-    "@angular/router": "~12.2.14",
-    "@kolkov/angular-editor": "^1.2.0",
-    "@ngx-translate/core": "^13.0.0",
-    "@types/nouislider": "^9.0.7",
+    "@angular-material-components/datetime-picker": "^9.0.0",
+    "@angular/animations": "^15.2.3",
+    "@angular/cdk": "^15.2.3",
+    "@angular/common": "^15.2.3",
+    "@angular/compiler": "^15.2.3",
+    "@angular/core": "^15.2.3",
+    "@angular/forms": "^15.2.3",
+    "@angular/material": "^15.2.3",
+    "@angular/platform-browser": "^15.2.3",
+    "@angular/platform-browser-dynamic": "^15.2.3",
+    "@angular/router": "^15.2.3",
+    "@kolkov/angular-editor": "3.0.0-beta.0",
+    "@ngx-translate/core": "^14.0.0",
     "lodash": "^4.17.20",
-    "memo-decorator": "^2.0.1",
-    "moment": "^2.29.0",
-    "ngx-markdown": "^12.0.1",
-    "ngx-mask": "12.0.0",
-    "nouislider": "^14.7.0",
-    "rxjs": "~6.6.7",
+    "moment": "^2.29.4",
+    "ngx-markdown": "^15.1.2",
+    "ngx-mask": "15.1.0",
+    "nouislider": "^15.7.0",
+    "prismjs": "^1.29.0",
+    "rxjs": "~7.8.0",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.13",
-    "@angular/cli": "~12.2.13",
-    "@angular/compiler-cli": "~12.2.14",
-    "@angular/language-service": "~12.2.14",
+    "@angular-devkit/build-angular": "^15.2.4",
+    "@angular-eslint/builder": "15.2.1",
+    "@angular-eslint/eslint-plugin": "15.2.1",
+    "@angular-eslint/eslint-plugin-template": "15.2.1",
+    "@angular-eslint/schematics": "15.2.1",
+    "@angular-eslint/template-parser": "15.2.1",
+    "@angular/cli": "^15.2.4",
+    "@angular/compiler-cli": "^15.2.3",
+    "@angular/language-service": "^15.2.3",
     "@types/jasmine": "~3.10.2",
     "@types/jasminewd2": "~2.0.10",
     "@types/lodash": "^4.14.178",
-    "@types/node": "^12.11.1",
-    "angular-cli-ghpages": "^1.0.0-rc.2",
-    "codelyzer": "^6.0.0",
+    "@types/node": "^18.15.5",
+    "@types/nouislider": "^15.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.43.0",
+    "@typescript-eslint/parser": "^5.43.0",
+    "angular-cli-ghpages": "^1.0.5",
     "cz-conventional-changelog": "^3.3.0",
+    "eslint": "^8.28.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^3.4.0",
     "husky": "^4.3.8",
     "jasmine-core": "~3.10.1",
     "jasmine-spec-reporter": "~7.0.0",
-    "karma": "~6.3.2",
+    "karma": "~6.4.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "lint-staged": "^10.5.4",
-    "ng-packagr": "^12.2.5",
+    "ng-packagr": "^15.2.2",
     "prettier": "2.3.1",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "tslint-plugin-prettier": "^2.3.0",
-    "typescript": "~4.3.5"
+    "typescript": "~4.9.5"
   }
 }
