@@ -1,11 +1,9 @@
-import { Component, ElementRef, Input, OnDestroy, ViewChild } from '@angular/core';
+import { AfterViewInit, Component, Input } from '@angular/core';
 import { TableCell } from '../../models/table-cell.model';
-import { combineLatest, Observable, ReplaySubject, Subscription } from 'rxjs';
-import { delay, filter, map, tap, shareReplay } from 'rxjs/operators';
+import { combineLatest, Observable, ReplaySubject } from 'rxjs';
+import { filter, map, shareReplay } from 'rxjs/operators';
 import { TooltipPosition } from '@angular/material/tooltip';
 import { readPropValue } from '../../../utils/utils';
-import { AfterViewInit } from '@angular/core/core';
-import { BehaviorSubject } from 'rxjs';
 
 function cellFormatter(cell: TableCell, data: any): string {
   if (cell?.cellFormatter) {
@@ -25,7 +23,7 @@ function cellFormatter(cell: TableCell, data: any): string {
   selector: 'lab900-table-cell-value[cell][data]',
   templateUrl: './table-cell-value.component.html',
 })
-export class Lab900TableCellValueComponent<T = any> implements OnDestroy, AfterViewInit {
+export class Lab900TableCellValueComponent<T = any> implements AfterViewInit {
   @Input()
   public set cell(value: TableCell<T>) {
     this._cell$.next(value);
@@ -35,11 +33,16 @@ export class Lab900TableCellValueComponent<T = any> implements OnDestroy, AfterV
   public set data(value: T) {
     this._data$.next(value);
   }
+
   private readonly _cell$ = new ReplaySubject<TableCell<T>>();
   private readonly _data$ = new ReplaySubject<T>();
 
-  public readonly cell$: Observable<TableCell<T>> = this._cell$.asObservable().pipe(shareReplay(1));
-  public readonly data$: Observable<T> = this._data$.asObservable().pipe(shareReplay(1));
+  public readonly cell$: Observable<TableCell<T>> = this._cell$
+    .asObservable()
+    .pipe(shareReplay(1));
+  public readonly data$: Observable<T> = this._data$
+    .asObservable()
+    .pipe(shareReplay(1));
 
   public readonly cellValue$: Observable<string>;
   public readonly icon$: Observable<string | null>;
@@ -57,22 +60,34 @@ export class Lab900TableCellValueComponent<T = any> implements OnDestroy, AfterV
     const streams = combineLatest([this.cell$, this.data$]);
     this.icon$ = streams.pipe(
       filter(([cell]) => !!cell?.icon),
-      map(([cell, data]) => cell.icon(data, cell)),
+      map(([cell, data]) => cell.icon(data, cell))
     );
     this.svgIcon$ = streams.pipe(
       filter(([cell]) => !!cell?.svgIcon),
-      map(([cell, data]) => cell.svgIcon(data, cell)),
+      map(([cell, data]) => cell.svgIcon(data, cell))
     );
-    this.cellValue$ = streams.pipe(map(([cell, data]) => cellFormatter(cell, data)));
-    this.maxWidth$ = streams.pipe(map(([cell, data]) => this.maxWidth(cell, data)));
+    this.cellValue$ = streams.pipe(
+      map(([cell, data]) => cellFormatter(cell, data))
+    );
+    this.maxWidth$ = streams.pipe(
+      map(([cell, data]) => this.maxWidth(cell, data))
+    );
   }
 
   public ngAfterViewInit(): void {
-    this.tooltip$ = combineLatest([this.cell$, this.data$, this.cellValue$]).pipe(
+    this.tooltip$ = combineLatest([
+      this.cell$,
+      this.data$,
+      this.cellValue$,
+    ]).pipe(
       map(([cell, data, value]) => this.getTooltipContent(cell, data, value)),
-      filter((content) => !!content?.length),
+      filter((content) => !!content?.length)
     );
-    this.tooltipPosition$ = this.cell$.pipe(map((cell) => cell.cellTooltip?.tooltipOptions?.tooltipPosition ?? 'below'));
+    this.tooltipPosition$ = this.cell$.pipe(
+      map(
+        (cell) => cell.cellTooltip?.tooltipOptions?.tooltipPosition ?? 'below'
+      )
+    );
   }
 
   public checkWidth(event): void {
@@ -81,7 +96,8 @@ export class Lab900TableCellValueComponent<T = any> implements OnDestroy, AfterV
       this.viewTooltip = true;
     }
     if (this.viewTooltip == null) {
-      this.viewTooltip = el.clientWidth < el.scrollWidth || el.clientHeight < el.scrollHeight;
+      this.viewTooltip =
+        el.clientWidth < el.scrollWidth || el.clientHeight < el.scrollHeight;
     }
   }
 
@@ -89,10 +105,12 @@ export class Lab900TableCellValueComponent<T = any> implements OnDestroy, AfterV
     return readPropValue<T>(cellMaxWidth, data) ?? this.maxColumnWidthFromTable;
   }
 
-  public getTooltipContent({ cellTooltip }: TableCell<T>, data: T, value: string): string {
+  public getTooltipContent(
+    { cellTooltip }: TableCell<T>,
+    data: T,
+    value: string
+  ): string {
     this.onlyOnOverflow = !!cellTooltip?.onlyOnOverflow;
     return readPropValue<T>(cellTooltip?.text, data) ?? value;
   }
-
-  public ngOnDestroy(): void {}
 }
