@@ -9,7 +9,6 @@ import { MatCardModule } from '@angular/material/card';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatMenuModule } from '@angular/material/menu';
-import { FlexLayoutModule } from '@angular/flex-layout';
 import { Lab900SharingModule } from '../sharing/sharing.module';
 import { TranslateModule } from '@ngx-translate/core';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
@@ -17,14 +16,22 @@ import { MatPaginatorModule } from '@angular/material/paginator';
 import { Lab900ButtonModule } from '../button/button.module';
 
 @NgModule({
-  declarations: [Lab900DataListComponent, Lab900DataListItemComponent, Lab900DataListEmptyDirective, Lab900DataListItemInfoDirective],
-  exports: [Lab900DataListComponent, Lab900DataListEmptyDirective, Lab900DataListItemInfoDirective],
+  declarations: [
+    Lab900DataListComponent,
+    Lab900DataListItemComponent,
+    Lab900DataListEmptyDirective,
+    Lab900DataListItemInfoDirective,
+  ],
+  exports: [
+    Lab900DataListComponent,
+    Lab900DataListEmptyDirective,
+    Lab900DataListItemInfoDirective,
+  ],
   imports: [
     CommonModule,
     MatCardModule,
     MatButtonModule,
     MatIconModule,
-    FlexLayoutModule,
     TranslateModule,
     MatMenuModule,
     Lab900SharingModule,
