@@ -1,4 +1,10 @@
-import { Directive, EventEmitter, HostListener, Input, Output } from '@angular/core';
+import {
+  Directive,
+  EventEmitter,
+  HostListener,
+  Input,
+  Output,
+} from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { ConfirmationDialogComponent } from '../components/confirmation-dialog/confirmation-dialog.component';
 import { SubscriptionBasedDirective } from '../../common/directives/subscription-based.directive';
@@ -14,11 +20,12 @@ export class ConfirmationDialogDirective extends SubscriptionBasedDirective {
   @Output() public confirmed: EventEmitter<void> = new EventEmitter<void>();
   @Output() public cancelled: EventEmitter<void> = new EventEmitter<void>();
 
-  constructor(public dialog: MatDialog) {
+  public constructor(public dialog: MatDialog) {
     super();
   }
 
-  @HostListener('click') public onMouseEnter(): void {
+  @HostListener('click')
+  public onMouseEnter(): void {
     const dialog = this.dialog.open(ConfirmationDialogComponent, {
       data: {
         message: this.message,
