@@ -1,4 +1,10 @@
-import { Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core';
+import {
+  Component,
+  Input,
+  OnChanges,
+  OnInit,
+  SimpleChanges,
+} from '@angular/core';
 import { MergeObject } from '../../models/merge-object.model';
 import { MergeConfig, MergeConfigBase } from '../../models/merge-config.model';
 import * as _ from 'lodash';
@@ -31,8 +37,11 @@ export class Lab900MergerComponent<T> implements OnInit, OnChanges {
 
   public result: T;
 
-  constructor(iconRegistry: MatIconRegistry, sanitizer: DomSanitizer) {
-    iconRegistry.addSvgIconLiteral('merge', sanitizer.bypassSecurityTrustHtml(MergeIcon));
+  public constructor(iconRegistry: MatIconRegistry, sanitizer: DomSanitizer) {
+    iconRegistry.addSvgIconLiteral(
+      'merge',
+      sanitizer.bypassSecurityTrustHtml(MergeIcon)
+    );
   }
 
   public ngOnInit(): void {
@@ -54,7 +63,10 @@ export class Lab900MergerComponent<T> implements OnInit, OnChanges {
         } else {
           const baseValue = this.getBase(s.active)[s.attribute];
           if (s.combine) {
-            this.result[s.attribute] = [...this.result[s.attribute], ...baseValue];
+            this.result[s.attribute] = [
+              ...this.result[s.attribute],
+              ...baseValue,
+            ];
           } else {
             this.result[s.attribute] = baseValue;
           }
@@ -69,7 +81,10 @@ export class Lab900MergerComponent<T> implements OnInit, OnChanges {
     } else if (config?.nestedObject) {
       let different = false;
       for (let i = 0; i < config?.nestedObject.length && !different; i++) {
-        different = this.compareValues(config.nestedObject[i].attribute, config.attribute);
+        different = this.compareValues(
+          config.nestedObject[i].attribute,
+          config.attribute
+        );
       }
       return different;
     } else {
@@ -78,20 +93,32 @@ export class Lab900MergerComponent<T> implements OnInit, OnChanges {
   }
 
   private compareValues(attribute: string, parentAttribute?: string): boolean {
-    const leftValue = parentAttribute ? this.leftObject.data[parentAttribute][attribute] : this.leftObject.data[attribute];
-    const rightValue = parentAttribute ? this.rightObject.data[parentAttribute][attribute] : this.rightObject.data[attribute];
+    const leftValue = parentAttribute
+      ? this.leftObject.data[parentAttribute][attribute]
+      : this.leftObject.data[attribute];
+    const rightValue = parentAttribute
+      ? this.rightObject.data[parentAttribute][attribute]
+      : this.rightObject.data[attribute];
 
     return !_.isEqual(
-      Array.isArray(this.leftObject.data[attribute]) ? _.sortBy(leftValue) : leftValue,
-      Array.isArray(this.rightObject.data[attribute]) ? _.sortBy(rightValue) : rightValue,
+      Array.isArray(this.leftObject.data[attribute])
+        ? _.sortBy(leftValue)
+        : leftValue,
+      Array.isArray(this.rightObject.data[attribute])
+        ? _.sortBy(rightValue)
+        : rightValue
     );
   }
 
   private getBase(active = false): T {
     if (active) {
-      return this.selected === 'right' ? this.leftObject.data : this.rightObject.data;
+      return this.selected === 'right'
+        ? this.leftObject.data
+        : this.rightObject.data;
     }
-    return this.selected === 'right' ? this.rightObject.data : this.leftObject.data;
+    return this.selected === 'right'
+      ? this.rightObject.data
+      : this.leftObject.data;
   }
 
   public toggleActive(config: MergeConfig<T>, index: number): void {
