@@ -3,15 +3,18 @@ import { IsActiveMatchOptions, NavigationEnd, Router } from '@angular/router';
 import { NavItem } from '../../models/nav-item.model';
 import { Subscription } from 'rxjs';
 import { SubscriptionBasedDirective } from '../../../common/directives/subscription-based.directive';
-import { MediaObserver } from '@angular/flex-layout';
 import { filter } from 'rxjs/operators';
+import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 
 @Component({
   selector: 'lab900-nav-item',
   templateUrl: './nav-item.component.html',
   styleUrls: ['./nav-item.component.scss'],
 })
-export class NavItemComponent extends SubscriptionBasedDirective implements OnInit, OnDestroy {
+export class NavItemComponent
+  extends SubscriptionBasedDirective
+  implements OnInit, OnDestroy
+{
   private sub: Subscription;
 
   @Input()
@@ -27,7 +30,7 @@ export class NavItemComponent extends SubscriptionBasedDirective implements OnIn
   public readonly depth = 0;
 
   @Input()
-  public readonly allowOverlayMenuUntil: string | string[] = 'xs';
+  public readonly allowOverlayMenuUntil: string | string[] = Breakpoints.XSmall;
 
   @Input()
   public disabled = false;
@@ -38,7 +41,10 @@ export class NavItemComponent extends SubscriptionBasedDirective implements OnIn
   @Input()
   public navListMatchOptions: IsActiveMatchOptions;
 
-  public constructor(public readonly router: Router, public readonly mediaObserver: MediaObserver) {
+  public constructor(
+    public readonly router: Router,
+    public readonly breakpointObserver: BreakpointObserver
+  ) {
     super();
   }
 
@@ -46,13 +52,18 @@ export class NavItemComponent extends SubscriptionBasedDirective implements OnIn
     if (!(this.item.route || this.item.href || this.item.children)) {
       this.disabled = true;
     } else {
-      this.addSubscription(this.router.events.pipe(filter((e) => e instanceof NavigationEnd)), (event: NavigationEnd) => {
-        this.item.navigationFinished?.(true);
-        const url = event.urlAfterRedirects;
-        if (url && this.item?.children?.length) {
-          this.expanded = this.item.children.some((item: NavItem) => url.indexOf(`/${item.route}`) === 0);
+      this.addSubscription(
+        this.router.events.pipe(filter((e) => e instanceof NavigationEnd)),
+        (event: NavigationEnd) => {
+          this.item.navigationFinished?.(true);
+          const url = event.urlAfterRedirects;
+          if (url && this.item?.children?.length) {
+            this.expanded = this.item.children.some(
+              (item: NavItem) => url.indexOf(`/${item.route}`) === 0
+            );
+          }
         }
-      });
+      );
     }
   }
 
@@ -74,7 +85,7 @@ export class NavItemComponent extends SubscriptionBasedDirective implements OnIn
       this.router.createUrlTree([item.route ?? item.href?.url], {
         queryParams: item.routeQueryParams,
       }),
-      item.routeMatchOptions ?? this.navListMatchOptions,
+      item.routeMatchOptions ?? this.navListMatchOptions
     );
   }
 }
