@@ -18,7 +18,6 @@ import { ThemePalette } from '@angular/material/core';
 import { MatCheckbox } from '@angular/material/checkbox';
 import { MatColumnDef, MatTable } from '@angular/material/table';
 import { SelectionModel } from '@angular/cdk/collections';
-import memo from 'memo-decorator';
 import { map } from 'rxjs/operators';
 
 @Component({
@@ -26,7 +25,9 @@ import { map } from 'rxjs/operators';
   templateUrl: './table-cell-select.component.html',
   encapsulation: ViewEncapsulation.None,
 })
-export class TableCellSelectComponent<T extends object = object> implements OnInit, OnDestroy {
+export class TableCellSelectComponent<T extends object = object>
+  implements OnInit, OnDestroy
+{
   @ViewChildren('rowCheckbox')
   public rowCheckboxes!: QueryList<MatCheckbox>;
 
@@ -46,9 +47,14 @@ export class TableCellSelectComponent<T extends object = object> implements OnIn
   public allSelected$: Observable<boolean>;
 
   @Input()
-  set selection(value: SelectionModel<T>) {
+  public set selection(value: SelectionModel<T>) {
     this._selection = value;
-    this.allSelected$ = this._selection.changed.pipe(map((change) => change?.source?.selected?.length === this.rowCheckboxes?.length));
+    this.allSelected$ = this._selection.changed.pipe(
+      map(
+        (change) =>
+          change?.source?.selected?.length === this.rowCheckboxes?.length
+      )
+    );
   }
 
   @Input()
@@ -81,16 +87,16 @@ export class TableCellSelectComponent<T extends object = object> implements OnIn
   }
 
   public handleSelectAllCheckbox({ checked }): void {
-    this.rowCheckboxes.toArray().forEach((checkBox) => (checkBox.checked = checked));
+    this.rowCheckboxes
+      .toArray()
+      .forEach((checkBox) => (checkBox.checked = checked));
     this.selectAll.emit(checked);
   }
 
-  @memo()
   public isChecked(value: T): boolean {
     return this._selection?.isSelected(value);
   }
 
-  @memo()
   public isDisabled(value: T): boolean {
     return this.isChecked(value) ? false : this.disabled;
   }
