@@ -1,7 +1,7 @@
 import { Directive } from '@angular/core';
 
 @Directive({
-  // tslint:disable-next-line:directive-selector
+  // eslint-disable-next-line @angular-eslint/directive-selector
   selector: '[lab900TableLeftFooter]',
 })
 export class Lab900TableLeftFooterDirective {}
