@@ -12,16 +12,16 @@ export class MarkdownPageComponent extends SubscriptionBasedDirective {
   @Input()
   public filePath: string;
 
-  constructor(private activatedRoute: ActivatedRoute) {
+  public constructor(private activatedRoute: ActivatedRoute) {
     super();
     this.addSubscription(
       this.activatedRoute.data.pipe(
         filter((data: { filePath: string }) => !!data?.filePath),
-        take(1),
+        take(1)
       ),
       (data) => {
         this.filePath = data.filePath;
-      },
+      }
     );
   }
 }
