@@ -25,11 +25,16 @@ export class ShowcasePageComponent extends SubscriptionBasedDirective {
   public data: ShowcaseRouteData;
   public navItems: PageHeaderNavItem[] = [];
 
-  constructor(private activatedRoute: ActivatedRoute, private router: Router) {
+  public constructor(
+    private activatedRoute: ActivatedRoute,
+    private router: Router
+  ) {
     super();
     this.addSubscription(this.activatedRoute.queryParams, (queryParams) => {
       this.data = this.activatedRoute.snapshot.data as ShowcaseRouteData;
-      this.navItems = !this.data?.docFile ? [this.exampleNav] : [this.guideNav, this.exampleNav];
+      this.navItems = !this.data?.docFile
+        ? [this.exampleNav]
+        : [this.guideNav, this.exampleNav];
       if (queryParams?.tab) {
         this.currentTab = queryParams?.tab;
       } else {
