@@ -1,4 +1,12 @@
-import { ComponentFactoryResolver, Directive, Input, OnChanges, SimpleChanges, Type, ViewContainerRef } from '@angular/core';
+import {
+  ComponentFactoryResolver,
+  Directive,
+  Input,
+  OnChanges,
+  SimpleChanges,
+  Type,
+  ViewContainerRef,
+} from '@angular/core';
 
 @Directive({
   selector: '[lab900ComponentLoader]',
@@ -7,7 +15,10 @@ export class ComponentLoaderDirective implements OnChanges {
   @Input()
   private lab900ComponentLoader: Type<any>;
 
-  public constructor(public container: ViewContainerRef, private resolver: ComponentFactoryResolver) {}
+  public constructor(
+    public container: ViewContainerRef,
+    private resolver: ComponentFactoryResolver
+  ) {}
 
   public ngOnChanges(changes: SimpleChanges): void {
     if (changes.lab900ComponentLoader) {
@@ -18,7 +29,9 @@ export class ComponentLoaderDirective implements OnChanges {
   private loadComponent(): void {
     this.container.clear();
     if (this.lab900ComponentLoader) {
-      const component = this.resolver.resolveComponentFactory(this.lab900ComponentLoader);
+      const component = this.resolver.resolveComponentFactory(
+        this.lab900ComponentLoader
+      );
       this.container.createComponent(component);
     }
   }
