@@ -13,7 +13,12 @@ export class ShowcaseRoute implements Route {
   public component: Type<any> = ShowcasePageComponent;
   public data: ShowcaseRouteData;
 
-  public constructor(public path: string, title: string, examples?: ShowcaseExample[], docFile?: string) {
+  public constructor(
+    public path: string,
+    title: string,
+    examples?: ShowcaseExample[],
+    docFile?: string
+  ) {
     this.data = { title, path, examples, docFile };
   }
 }
